Various fixes to defeat g++ compiler warnings, removal of some deprecated properties/classes

+++ src/osgEarthFeatures/FeatureModelGraph.cpp
@@ -323,7 +323,7 @@ FeatureModelGraph::load( unsigned lod, unsigned tileX, unsigned tileY, const std
         osg::Group* geometry =0L;
         const FeatureProfile* featureProfile = _session->getFeatureSource()->getFeatureProfile();
 
-        if ( lod >= featureProfile->getFirstLevel() )
+        if ( (int)lod >= featureProfile->getFirstLevel() )
         {
             // The extent of this tile:
             GeoExtent tileExtent = s_getTileExtent( lod, tileX, tileY, _usableFeatureExtent );
@@ -347,7 +347,7 @@ FeatureModelGraph::load( unsigned lod, unsigned tileX, unsigned tileY, const std
             result = geometry;
         }
 
-        if ( lod < featureProfile->getMaxLevel() )
+        if ( (int)lod < featureProfile->getMaxLevel() )
         {
             // see if there are any more levels. If so, build some pagedlods to bring the
             // next level in.
@@ -367,7 +367,7 @@ FeatureModelGraph::load( unsigned lod, unsigned tileX, unsigned tileY, const std
                 //}
 
                 // only build sub-pagedlods if we are expecting subtiles at some point:
-                if ( geometry != 0L || lod < featureProfile->getFirstLevel() )
+                if ( geometry != 0L || (int)lod < featureProfile->getFirstLevel() )
                 {
                     MapFrame mapf = _session->createMapFrame();
                     buildSubTilePagedLODs( lod, tileX, tileY, &mapf, group.get() );
@@ -392,7 +392,7 @@ FeatureModelGraph::load( unsigned lod, unsigned tileX, unsigned tileY, const std
         result = build( all, GeoExtent::INVALID, 0 );
     }
 
-    else if ( lod < _lodmap.size() )
+    else if ( (int)lod < _lodmap.size() )
     {
         // This path computes the SG for a model graph with explicity-defined levels of
         // detail. We already calculated the LOD level map in setupPaging(). If the
